+++ src/main/webapp/resources/js/contexts/roles-context.js
@@ -1,8 +1,6 @@
-import React, { createContext, useEffect, useState } from "react";
-import { getProjectRoles } from "../apis/projects/projects";
+import React, { createContext, useContext, useEffect, useState } from "react";
 
-let ProjectRolesContext;
-const { Provider } = (ProjectRolesContext = createContext());
+const RolesContext = createContext();
 
 /**
  * React context to allow multiple components to get a list of project roles
@@ -12,15 +10,15 @@ const { Provider } = (ProjectRolesContext = createContext());
  * @returns {*}
  * @constructor
  */
-function ProjectRolesProvider({ children }) {
+function RolesProvider({ children, rolesFn }) {
   const [roles, setRoles] = useState([]);
 
   /*
   When the component is mounted, get the list up current roles from the server.
    */
   useEffect(() => {
-    getProjectRoles().then((data) => setRoles(data));
-  }, []);
+    rolesFn().then((data) => setRoles(data));
+  }, [rolesFn]);
 
   /**
    * Find the translation for any project role.  If the role is not found,
@@ -34,7 +32,19 @@ function ProjectRolesProvider({ children }) {
     return role ? role.label : "UNKNOWN";
   };
 
-  return <Provider value={{ roles, getRoleFromKey }}>{children}</Provider>;
+  return (
+    <RolesContext.Provider value={{ roles, getRoleFromKey }}>
+      {children}
+    </RolesContext.Provider>
+  );
 }
 
-export { ProjectRolesContext, ProjectRolesProvider };
+function useRoles() {
+  const context = useContext(RolesContext);
+  if (context === undefined) {
+    throw new Error("useRoles must be used within a RolesProvider");
+  }
+  return context;
+}
+
+export { RolesProvider, useRoles };